3000mod/Builds/build3_functional/requirements.txt
pyserial>=3.5
PyYAML>=6.0
pytest>=7.0.0
